Refactor models for VC issuance
