+++ app/src/main/java/by/alexandr7035/affinidi_id/presentation/issue_credential/AvailableCredentialsAdapter.kt
@@ -37,7 +37,7 @@ class AvailableCredentialsAdapter(private val credentialClickListener: Credentia
             binding.issuerView.text = item.issuer
 
             binding.root.setOnClickListener {
-                credentialClickListener.onClick(item.vcType)
+                credentialClickListener.onClick(item.issuingCredentialType)
             }
         }
     }
