+++ app/src/main/java/by/alexandr7035/affinidi_id/presentation/issue_credential/CredentialClickListener.kt
@@ -1,7 +1,7 @@
 package by.alexandr7035.affinidi_id.presentation.issue_credential
 
-import by.alexandr7035.affinidi_id.domain.model.credentials.common.VcType
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssuingCredentialType
 
 interface CredentialClickListener {
-    fun onClick(availableVcType: VcType)
+    fun onClick(availableIssuingCredentialType: IssuingCredentialType)
 }
