+++ data/src/main/java/by/alexandr7035/data/helpers/vc_mapping/CredentialSubjectCaster.kt
@@ -1,6 +1,6 @@
 package by.alexandr7035.data.helpers.vc_mapping
 
-import by.alexandr7035.affinidi_id.domain.model.credentials.common.credential_subject.CredentialSubjectData
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.credential_subject.CredentialSubjectData
 import by.alexandr7035.data.model.CredentialSubject
 
 interface CredentialSubjectCaster {
