+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/available_credential_types/AvailableCredentialTypeModel.kt
@@ -1,10 +1,10 @@
 package by.alexandr7035.affinidi_id.domain.model.credentials.available_credential_types
 
-import by.alexandr7035.affinidi_id.domain.model.credentials.common.VcType
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssuingCredentialType
 
 data class AvailableCredentialTypeModel(
     val typeName: String,
     val description: String,
     val issuer: String,
-    val vcType: VcType
+    val issuingCredentialType: IssuingCredentialType
 )
