--- domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/common/VcType.kt
@@ -1,6 +1,0 @@
-package by.alexandr7035.affinidi_id.domain.model.credentials.common
-
-enum class VcType {
-    EMAIL_CREDENTIAL,
-    UNKNOWN_CREDENTIAL
-}
