--- domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/common/credential_subject/CredentialSubjectData.kt
@@ -1,3 +1,0 @@
-package by.alexandr7035.affinidi_id.domain.model.credentials.common.credential_subject
-
-abstract class CredentialSubjectData
