--- domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/common/credential_subject/EmailCredentialSubjectData.kt
@@ -1,5 +1,0 @@
-package by.alexandr7035.affinidi_id.domain.model.credentials.common.credential_subject
-
-data class EmailCredentialSubjectData(
-    val email: String
-): CredentialSubjectData()
