--- domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/common/credential_subject/UnknownCredentialSubjectData.kt
@@ -1,3 +1,0 @@
-package by.alexandr7035.affinidi_id.domain.model.credentials.common.credential_subject
-
-class UnknownCredentialSubjectData: CredentialSubjectData()
