+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/issue_vc/IssuingCredentialType.kt
@@ -1,0 +1,5 @@
+package by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc
+
+enum class IssuingCredentialType {
+    EMAIL_CREDENTIAL
+}
