+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/issue_vc/credential_subject/CredentialSubjectData.kt
@@ -1,0 +1,3 @@
+package by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.credential_subject
+
+abstract class CredentialSubjectData
