+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/model/credentials/issue_vc/credential_subject/EmailCredentialSubjectData.kt
@@ -1,0 +1,5 @@
+package by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.credential_subject
+
+data class EmailCredentialSubjectData(
+    val email: String
+): CredentialSubjectData()
