+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/usecase/credentials/GetAvailableVcTypesUseCase.kt
@@ -1,7 +1,7 @@
 package by.alexandr7035.affinidi_id.domain.usecase.credentials
 
 import by.alexandr7035.affinidi_id.domain.model.credentials.available_credential_types.AvailableCredentialTypeModel
-import by.alexandr7035.affinidi_id.domain.model.credentials.common.VcType
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssuingCredentialType
 
 class GetAvailableVcTypesUseCase {
     fun execute(): List<AvailableCredentialTypeModel> {
@@ -11,7 +11,7 @@ class GetAvailableVcTypesUseCase {
                 typeName = "EmailCredential",
                 description = "This VC proves the ownership of e-mail address used as login in AffinidiID.",
                 issuer = "self-claimed",
-                vcType = VcType.EMAIL_CREDENTIAL
+                issuingCredentialType = IssuingCredentialType.EMAIL_CREDENTIAL
             )
         )
     }
